Add shortcut to search in specs
specs are not only in the `spec/` dir. (This is related to the packwerk
gem.) It means that there are lots of `spec/` dirs. That makes my
searches trickier since I can't just do `:Rg search_term spec/`. The
solution is `:Rg -g '*_spec.rb' search_term`, but that's a little tricky
so `<leader>s` is a lot easier.